test(patina): keep casing option sources bounded
